Organize routers; seperate into controllers; make testing file and folder; npm install jest
