+++ package.json
@@ -1,5 +1,8 @@
 {
   "dependencies": {
     "dotenv": "^16.4.5"
+  },
+  "devDependencies": {
+    "jest": "^29.7.0"
   }
 }
